fix: only use locator class for accountid
