+++ run/test/specs/utils/get_account_id.ts
@@ -1,17 +1,6 @@
-import { DeviceWrapper } from '../../../types/DeviceWrapper';
 import { User } from '../../../types/testing';
 import { SupportedPlatformsType } from './open_app';
 
-export const saveSessionIdIos = async (device: DeviceWrapper) => {
-  const selector = await device.grabTextFromAccessibilityId('Session ID generated');
-  return selector;
-};
-
-export const getAccountId = async (device: DeviceWrapper) => {
-  const AccountId = await device.grabTextFromAccessibilityId('Account ID');
-
-  return AccountId;
-};
 
 export function sortByPubkey(...users: Array<User>) {
   return [...users]
