+++ run/test/specs/utils/index.ts
@@ -1,12 +1,9 @@
 import { clickOnCoordinates } from './click_by_coordinates';
-import { getAccountId, saveSessionIdIos } from './get_account_id';
 import { runOnlyOnAndroid, runOnlyOnIOS } from './run_on';
 import { sleepFor } from './sleep_for';
 
 export {
   sleepFor,
-  saveSessionIdIos,
-  getAccountId,
   runOnlyOnIOS,
   runOnlyOnAndroid,
   clickOnCoordinates,
